Fix compile warning

+++ src/mode/mixins/bulletin_drawer.cpp
@@ -130,7 +130,7 @@ void BulletinDrawer::bulletin_draw_text() {
 }
 
 void BulletinDrawer::bulletin_war_ally_init() {
-    for (int i = 0; i < sizeof(bulletin_war_ally_ids_army); ++i) {
+    for (long unsigned int i = 0; i < sizeof(bulletin_war_ally_ids_army); ++i) {
         bulletin_war_ally_ids_army[i] = draw_manager.new_sprite_id();
     }
     bulletin_war_ally_id_exit = draw_manager.new_sprite_id();
@@ -183,7 +183,7 @@ void BulletinDrawer::bulletin_war_ally_init() {
 
 void BulletinDrawer::bulletin_war_ally_update() {
     if (draw_manager.query_click(bulletin_war_ally_id_exit).id) {
-        for (int i = 0; i < sizeof(bulletin_war_ally_ids_army); ++i) {
+        for (long unsigned int i = 0; i < sizeof(bulletin_war_ally_ids_army); ++i) {
             draw_manager.draw(bulletin_war_ally_ids_army[i], nullptr);
             draw_manager.release_sprite_id(bulletin_war_ally_ids_army[i]);
         }
